Move user related stuff to user.py
